+++ bespokeasm/assembler/model/instruction_set.py
@@ -18,6 +18,8 @@ class InstructionSet(dict[str, InstructionBase]):
             ):
         self._instructions_config = instructions_config
         self._macros_config = macros_config
+        self._instruction_mnemonics: set[str] = set()
+        self._macro_mnemonics: set[str] = set()
 
         lower_keywords = {kw.lower(): kw for kw in ASSEMBLER_KEYWORD_SET}
 
@@ -26,6 +28,7 @@ class InstructionSet(dict[str, InstructionBase]):
             if mnemonic in lower_keywords:
                 sys.exit(f'ERROR - ISA configuration defined instruction "{mnemonic}" which is also a BespokeASM keyword')
             self[mnemonic] = Instruction(mnemonic, instr_config, operand_set_collection, default_endian, registers)
+            self._instruction_mnemonics.add(mnemonic)
 
         if self._macros_config is not None:
             # build a list of macros first so a macro can't be defined based on other macros
@@ -42,6 +45,15 @@ class InstructionSet(dict[str, InstructionBase]):
                 )
             for macro in macro_list:
                 self[macro.mnemonic] = macro
+                self._macro_mnemonics.add(macro.mnemonic)
 
     def get(self, mnemonic: str) -> InstructionBase:
         return super().get(mnemonic, None)
+
+    @property
+    def instruction_mnemonics(self) -> set[str]:
+        return self._instruction_mnemonics
+
+    @property
+    def macro_mnemonics(self) -> set[str]:
+        return self._macro_mnemonics